Updated swingx to 1.6.1

+++ application/trunk/baralga-core/src/main/java/org/remast/baralga/gui/panels/table/HoursByDayTableFormat.java
@@ -41,7 +41,7 @@ public class HoursByDayTableFormat implements TableFormat<HoursByDay> {
         case 0:
             return baseObject.getDay();
         case 1:
-            return baseObject.getHours();
+            return (Double) baseObject.getHours();
         default:
             return null;
         }
